+++ module-apps/application-settings-new/windows/BluetoothWindow.hpp
@@ -11,14 +11,14 @@ namespace gui
     {
       public:
         BluetoothWindow(app::Application *app);
-        ~BluetoothWindow() override = default;
         void onBeforeShow(ShowMode mode, SwitchData *data) override;
 
       private:
         bool isBluetoothSwitchOn       = false;
         bool isPhoneVisibilitySwitchOn = false;
         auto bluetoothOptionsList() -> std::list<gui::Option>;
-        void switchHandler(bool &switchState);
+        void bluetoothSwitchHandler(bool &switchState);
+        void phoneVisibilitySwitchHandler(bool &switchState);
         void rebuildOptionList();
     };
 }; // namespace gui